Add script to run new GWLF-E once at a time
Since the performance optimizations for GWLF-E rely
on memoization techniques, running multiple files on
the same Parser instance results in unstable output.

By making a new test that runs only one file at a time,
we test the new GWLF-E code on more reliable ground.

It remains to be seen if the multirun functionality is
truly required by MMW or not.